pull out capabilities extension and make it compliant
